separated components recipe-list & recipe-item; they now communicate to each other
