--- first-project/.log/ti-16272.log
@@ -1,19 +1,0 @@
-[19:03:08.159] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[19:03:08.160] Processing cache location '/home/albert/.cache/typescript/4.7'
-[19:03:08.160] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[19:03:08.161] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[19:03:08.161] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[19:03:08.161] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[19:03:08.161] Process id: 16279
-[19:03:08.161] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[19:03:08.161] validateDefaultNpmLocation: false
-[19:03:08.161] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[19:03:08.162] Updating types-registry npm package...
-[19:03:08.162] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[19:03:09.014]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 624ms
-
-found 0 vulnerabilities
-
-[19:03:09.014] Updated types-registry npm package
