--- first-project/.log/ti-4456.log
@@ -1,20 +1,0 @@
-[14:06:32.927] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[14:06:32.928] Processing cache location '/home/albert/.cache/typescript/4.7'
-[14:06:32.928] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[14:06:32.930] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[14:06:32.930] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[14:06:32.930] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[14:06:32.930] Process id: 4463
-[14:06:32.930] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[14:06:32.930] validateDefaultNpmLocation: false
-[14:06:32.931] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[14:06:32.931] Updating types-registry npm package...
-[14:06:32.931] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[14:06:33.878]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 726ms
-
-found 0 vulnerabilities
-
-[14:06:33.878] Updated types-registry npm package
-[14:21:57.656] Parent process has exited, shutting down...
