--- first-project/.log/ti-5617.log
@@ -1,19 +1,0 @@
-[14:22:32.514] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[14:22:32.514] Processing cache location '/home/albert/.cache/typescript/4.7'
-[14:22:32.514] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[14:22:32.515] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[14:22:32.515] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[14:22:32.515] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[14:22:32.515] Process id: 5624
-[14:22:32.515] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[14:22:32.515] validateDefaultNpmLocation: false
-[14:22:32.516] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[14:22:32.516] Updating types-registry npm package...
-[14:22:32.516] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[14:22:33.368]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 574ms
-
-found 0 vulnerabilities
-
-[14:22:33.368] Updated types-registry npm package
