--- first-project/.log/ti-6923.log
@@ -1,35 +1,0 @@
-[14:32:12.884] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[14:32:12.885] Processing cache location '/home/albert/.cache/typescript/4.7'
-[14:32:12.885] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[14:32:12.885] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[14:32:12.885] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[14:32:12.886] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[14:32:12.886] Process id: 6930
-[14:32:12.886] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[14:32:12.886] validateDefaultNpmLocation: false
-[14:32:12.886] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[14:32:12.886] Updating types-registry npm package...
-[14:32:12.886] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[14:32:13.805]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 691ms
-
-found 0 vulnerabilities
-
-[14:32:13.805] Updated types-registry npm package
-[15:33:39.089] Got install request {"projectName":"/dev/null/inferredProject1*","fileNames":[],"compilerOptions":{"module":1,"target":3,"jsx":1,"allowJs":true,"allowSyntheticDefaultImports":true,"allowNonTsExtensions":true,"noEmitForJsFiles":true},"typeAcquisition":{"enable":true,"include":[],"exclude":[]},"unresolvedImports":[],"projectRootPath":"/home/albert/Documentos/courses/angular/first-project/src/app/recipes","kind":"discover"}
-[15:33:39.133] Loaded safelist from types map file '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json'
-[15:33:39.134] Explicitly included types: []
-[15:33:39.136] Inferred typings from unresolved imports: []
-[15:33:39.136] Result: {"cachedTypingPaths":[],"newTypingNames":[],"filesToWatch":["/home/albert/Documentos/courses/angular/first-project/src/app/recipes/bower_components","/home/albert/Documentos/courses/angular/first-project/src/app/recipes/node_modules"]}
-[15:33:39.136] Finished typings discovery: {"cachedTypingPaths":[],"newTypingNames":[],"filesToWatch":["/home/albert/Documentos/courses/angular/first-project/src/app/recipes/bower_components","/home/albert/Documentos/courses/angular/first-project/src/app/recipes/node_modules"]}
-[15:33:39.137] DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/bower_components
-[15:33:39.138] DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/bower_components 1 undefined Project: /dev/null/inferredProject1* watcher already invoked: false
-[15:33:39.142] Elapsed:: 3.0951630007475615ms DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/bower_components 1 undefined Project: /dev/null/inferredProject1* watcher already invoked: false
-[15:33:39.142] DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/node_modules
-[15:33:39.142] DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/node_modules 1 undefined Project: /dev/null/inferredProject1* watcher already invoked: false
-[15:33:39.142] Elapsed:: 0.1896160002797842ms DirectoryWatcher:: Added:: WatchInfo: /home/albert/Documentos/courses/angular/first-project/src/app/recipes/node_modules 1 undefined Project: /dev/null/inferredProject1* watcher already invoked: false
-[15:33:39.142] Sending response:
-    {"projectName":"/dev/null/inferredProject1*","typeAcquisition":{"enable":true,"include":[],"exclude":[]},"compilerOptions":{"module":1,"target":3,"jsx":1,"allowJs":true,"allowSyntheticDefaultImports":true,"allowNonTsExtensions":true,"noEmitForJsFiles":true},"typings":[],"unresolvedImports":[],"kind":"action::set"}
-[15:33:39.143] Response has been sent.
-[15:33:39.143] No new typings were requested as a result of typings discovery
