--- first-project/.log/ti-7939.log
@@ -1,19 +1,0 @@
-[17:23:23.949] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[17:23:23.950] Processing cache location '/home/albert/.cache/typescript/4.7'
-[17:23:23.950] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[17:23:23.951] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[17:23:23.951] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[17:23:23.951] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[17:23:23.951] Process id: 7946
-[17:23:23.951] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[17:23:23.951] validateDefaultNpmLocation: false
-[17:23:23.951] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[17:23:23.951] Updating types-registry npm package...
-[17:23:23.951] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[17:23:24.767]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 532ms
-
-found 0 vulnerabilities
-
-[17:23:24.767] Updated types-registry npm package
