--- first-project/.log/ti-8630.log
@@ -1,19 +1,0 @@
-[15:37:42.745] Global cache location '/home/albert/.cache/typescript/4.7', safe file path '/home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typingSafeList.json', types map path /home/albert/Documentos/courses/angular/first-project/node_modules/typescript/lib/typesMap.json
-[15:37:42.746] Processing cache location '/home/albert/.cache/typescript/4.7'
-[15:37:42.746] Trying to find '/home/albert/.cache/typescript/4.7/package.json'...
-[15:37:42.746] Loaded content of '/home/albert/.cache/typescript/4.7/package.json': {"private":true,"dependencies":{"types-registry":"^0.1.599"}}
-[15:37:42.746] Loaded content of '/home/albert/.cache/typescript/4.7/package-lock.json'
-[15:37:42.746] Finished processing cache location '/home/albert/.cache/typescript/4.7'
-[15:37:42.747] Process id: 8637
-[15:37:42.747] NPM location: /usr/bin/npm (explicit '--npmLocation' not  provided)
-[15:37:42.747] validateDefaultNpmLocation: false
-[15:37:42.747] Npm config file: /home/albert/.cache/typescript/4.7/package.json
-[15:37:42.747] Updating types-registry npm package...
-[15:37:42.747] Exec: /usr/bin/npm install --ignore-scripts types-registry@latest
-[15:37:43.632]     Succeeded. stdout:
-    
-    up to date, audited 2 packages in 644ms
-
-found 0 vulnerabilities
-
-[15:37:43.632] Updated types-registry npm package
