+++ first-project/src/app/app.component.ts
@@ -7,4 +7,9 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   title = 'first-project';
+  loadedFeature = 'recipe';
+
+  onNavigate(featre: string) {
+    this.loadedFeature = featre;
+  }
 }
