+++ first-project/src/app/header/header.component.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { Component, EventEmitter, Output } from "@angular/core";
 
 @Component({
     selector: 'app-header',
@@ -6,5 +6,10 @@ import { Component } from "@angular/core";
 })
 
 export class HeaderComponent {
+    @Output() featureSelected = new EventEmitter<string>();
     collapsed = true;
+
+    onSelect(feature: string) {
+        this.featureSelected.emit(feature);
+    }
 }
