+++ first-project/src/app/recipes/recipe-list/recipe-item/recipe-item.component.html
@@ -1,0 +1,12 @@
+<a href="#" class="list-group-item clearfix">
+    <div class="pull-left">
+        <h4 class="list-group-item-heading">{{ recipe.name  }}</h4>
+        <p class="list-group-item-text">{{ recipe.description  }}</p>
+    </div>
+    <span class="pull-right">
+        <img [src]="recipe.imagePath"
+             alt="{{ recipe.name }}"
+             class="img-responsive"
+             style="max-height: 50px;"/>
+    </span>
+</a>
