+++ first-project/src/app/recipes/recipe-list/recipe-item/recipe-item.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { Recipe } from '../../recipe.model';
+
+@Component({
+  selector: 'app-recipe-item',
+  templateUrl: './recipe-item.component.html',
+  styleUrls: ['./recipe-item.component.css']
+})
+export class RecipeItemComponent implements OnInit {
+  @Input() recipe: Recipe;
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
