+++ first-project/src/app/recipes/recipe-list/recipe-list.component.html
@@ -6,18 +6,8 @@
 <hr>
 <div class="row">
     <div class="col-xs-12">
-        <a href="#" class="list-group-item clearfix" *ngFor="let recipe of recipes">
-            <div class="pull-left">
-                <h4 class="list-group-item-heading">{{ recipe.name  }}</h4>
-                <p class="list-group-item-text">{{ recipe.description  }}</p>
-            </div>
-            <span class="pull-right">
-                <img [src]="recipe.imagePath"
-                     alt="{{ recipe.name }}"
-                     class="img-responsive"
-                     style="max-height: 50px;"/>
-            </span>
-        </a>
-        <app-recipe-item></app-recipe-item>
+        <app-recipe-item
+            *ngFor="let recipeElement of recipes"
+            [recipe]="recipeElement"></app-recipe-item>
     </div>
 </div>
