+++ first-project/src/app/recipes/recipe-list/recipe-list.component.ts
@@ -9,6 +9,8 @@ import { Recipe } from '../recipe.model';
 export class RecipeListComponent implements OnInit {
   recipes: Recipe[] = [
     new Recipe('testRecipe', 'this is dummy',
+               'https://get.pxhere.com/photo/penne-dish-cuisine-italian-food-european-food-penne-alla-vodka-food-pasta-pomodoro-pasta-spaghetti-alla-puttanesca-vegetarian-food-side-dish-rigatoni-american-food-junk-food-recipe-1539171.jpg'),
+    new Recipe('testRecipe', 'this is dummy',
                'https://get.pxhere.com/photo/penne-dish-cuisine-italian-food-european-food-penne-alla-vodka-food-pasta-pomodoro-pasta-spaghetti-alla-puttanesca-vegetarian-food-side-dish-rigatoni-american-food-junk-food-recipe-1539171.jpg')
   ];
