--- first-project/src/app/recipes/recipes-list/recipe-item/recipe-item.component.html
@@ -1,1 +1,0 @@
-<p>recipe-item works!</p>
